Some changes:
* Remove PDF Chart
* Add pdf computing to QWS
* Add print if Java process execution return error
* Add Java properties to QWS
* Update the README files and add information about the input format

+++ QWS/src/QWS.java
@@ -20,6 +20,11 @@ package qws;
 import java.io.File;
 import java.io.FileReader;
 import java.io.BufferedReader;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.FileInputStream;
+import java.util.Properties;
+import java.io.IOException;
 
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.FileSystem;
@@ -40,52 +45,35 @@ import org.apache.hadoop.conf.Configuration;
 
 public class QWS {
 
-    /**
-     * The number of steps that will be executed in the simulation.
-     */
-    private static final int STEPS = 8;
-
-    /**
-     * The path of the file that contains the paths to the files of all matrices
-     * U and the vector psi. The order matter. Should be: U_0,U_1,...,U_n,psi.
-     * One per line.
-     */
-    private static final String FILES =
-            "/home/david/Desktop/java/QWS/input/files";
-
-    /**
-     * The path in the HDFS where the program will store the data.
-     */
-    private static final String WORK_DIR = "qws_tmp/";
-
-    /**
-     * The folder path where the .jar files are stored.
-     */
-    private static final String JAR_DIR = "/home/david/Desktop/java/QWS/";
-
-    /**
-     * The folder path where the result will be stored. Should be a empty folder
-	 * because this ALL DATA will be deleted.
-     */
-    private static final String OUTPUT_DIR =
-            "/home/david/Desktop/java/QWS/Result/";
-
-
     public static void main(String[] args) throws Exception {
 
         long startTime;
+        String steps;
+        String paths;
+        String workDir;
+        String jarDir;
+        String newShape;
+        String axes;
+        String position;
+        String outputDir;
         String line;
         String psi;
         String psiT;
         String psiTNorm;
+        String absSquare;
+        String reshape;
+        String pdf;
+        BufferedReader br;
         int numberU;
         String[] uDir;
         Runtime rt;
         Process pr;
+        Properties prop = new Properties();
+	    InputStream configInput = null;
 
         Configuration conf = new Configuration();
         FileSystem fs;
-        FileUtil fu;
+        FileUtil fu = new FileUtil();
         Path pt;
         FileStatus[] status;
 
@@ -94,17 +82,58 @@ public class QWS {
             startTime = System.nanoTime();
 
             fs = FileSystem.get(conf);
-            fu = new FileUtil();
+
+            configInput = new FileInputStream("config.properties");
+
+            // load the properties file
+		    prop.load(configInput);
+
+            // Get the configurations in the properties file
+		    steps = prop.getProperty("steps");
+		    paths = prop.getProperty("paths");
+		    workDir = prop.getProperty("workDir");
+		    jarDir = prop.getProperty("jarDir");
+		    newShape = prop.getProperty("newShape");
+		    axes = prop.getProperty("axes");
+		    position = prop.getProperty("position");
+		    outputDir = prop.getProperty("outputDir");
+
+            if (steps == null || steps.equals("")) {
+                throw new IOException ("The value of the configuration "
+                        + "\"steps\" can not be null or empty.");
+            }
+
+            if (paths == null || paths.equals("")) {
+                throw new IOException ("The value of the configuration "
+                        + "\"paths\" can not be null or empty.");
+            } 
+
+            if (workDir == null || workDir.equals("")) {
+                throw new IOException ("The value of the configuration "
+                        + "\"workDir\" can not be null or empty.");
+            }
+
+            if (jarDir == null || jarDir.equals("")) {
+                throw new IOException ("The value of the configuration "
+                        + "\"jarDir\" can not be null or empty.");
+            }
+
+            if (outputDir == null || outputDir.equals("")) {
+                throw new IOException ("The value of the configuration "
+                        + "\"outputDir\" can not be null or empty.");
+            }
+
+            configInput.close();
 
             /*
-             * Delete the WORK_DIR directory if exists. And create a new one
+             * Delete the workDir directory if exists. And create a new one
              * empty.
              */
-            pt = new Path(WORK_DIR);
+            pt = new Path(workDir);
             fs.delete(pt, true);
             fs.mkdirs(pt);
 
-            BufferedReader br = new BufferedReader(new FileReader(FILES));
+            br = new BufferedReader(new FileReader(paths));
 
             /*
              * The loop below computes the number of U passed in input and
@@ -123,7 +152,7 @@ public class QWS {
 
             uDir = new String[numberU];
 
-            br = new BufferedReader(new FileReader(FILES));
+            br = new BufferedReader(new FileReader(paths));
             line = br.readLine();
 
             psi = "";
@@ -154,28 +183,28 @@ public class QWS {
 
             br.close();
 
-            pt = new Path(WORK_DIR + "psi");
+            pt = new Path(workDir + "psi");
             fs.delete(pt, true);
             fs.copyFromLocalFile(new Path(psi), pt);
 
             for (int i = 0; i < numberU; i++) {
-                pt = new Path(WORK_DIR + "u" + Integer.toString(i));
+                pt = new Path(workDir + "u" + Integer.toString(i));
                 fs.delete(pt, true);
                 fs.copyFromLocalFile(new Path(uDir[i]), pt);
             }
 
-            psiT = WORK_DIR + "psiT";
+            psiT = workDir + "psiT";
             pt = new Path(psiT);
             fs.mkdirs(pt);
 
-            status = fs.listStatus(new Path(WORK_DIR + "psi"));
+            status = fs.listStatus(new Path(workDir + "psi"));
             for (FileStatus stat : status) {
 
                 fu.copy(fs, stat.getPath(), fs, pt, false, true, conf);
             }
 
             for (int i = 0; i < numberU; i++) {
-                pt = new Path(WORK_DIR + "u" + Integer.toString(i));
+                pt = new Path(workDir + "u" + Integer.toString(i));
                 status = fs.listStatus(pt);
                 for (FileStatus stat : status) {
 
@@ -191,7 +220,7 @@ public class QWS {
 
             rt = Runtime.getRuntime();
 
-            for (int i = 0; i < STEPS; i++) {
+            for (int i = 0; i < Integer.parseInt(steps); i++) {
 
                 for (int j = numberU - 1; j > -1; j--) {
 
@@ -203,18 +232,12 @@ public class QWS {
                     for (FileStatus stat : status) {
 
                         if (stat.getPath().toString().indexOf("_logs") > -1) {
-
-                            //fs.delete(stat.getPath(), true);
                             continue;
                         } else {
-
                             if (stat.getPath().toString().indexOf("_SUCCESS")
                                     > -1) {
-
-                                //fs.delete(stat.getPath(), true);
                                 continue;
                             } else {
-
                                 fs.rename(stat.getPath(), pt);
                             }
                         }
@@ -225,20 +248,32 @@ public class QWS {
 
                     pt = new Path(psiT + "_tmp");
 
-                    status = fs.listStatus(new Path(WORK_DIR + "u"
+                    status = fs.listStatus(new Path(workDir + "u"
                                 + Integer.toString(j)));
                     for (FileStatus stat : status) {
 
                         fu.copy(fs, stat.getPath(), fs, pt, false, true, conf);
                     }
 
-                    pr = rt.exec("hadoop jar " + JAR_DIR
+                    pr = rt.exec("hadoop jar " + jarDir
                             + "operations.jar operations.MultMatrix " + psiT
-                            + "_tmp" + " " + WORK_DIR + "tmp" + " " + psiT
+                            + "_tmp" + " " + workDir + "tmp" + " " + psiT
                             + " B");
 
-
                     pr.waitFor();
+
+                    br = new BufferedReader(new InputStreamReader(
+                            pr.getInputStream()));
+
+                    if (pr.exitValue() != 0) {
+
+                        while ((line = br.readLine()) != null) {
+                            System.out.println(line);
+                        }
+
+                        System.exit(1);
+                    }
+
                     pr.destroy();
 
                 }
@@ -247,20 +282,51 @@ public class QWS {
             }
 
 
-            psiTNorm = WORK_DIR + "psiTNorm";
+            psiTNorm = workDir + "psiTNorm";
 
             // Delete the output directory if exists.
             pt = new Path(psiTNorm);
             fs.delete(pt, true);
 
-            pr = rt.exec("hadoop jar " + JAR_DIR + "operations.jar operations."
+            pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
                     + "NormMatrix " + psiT + " " + psiTNorm);
 
             pr.waitFor();
+
+            br = new BufferedReader(new InputStreamReader(
+                    pr.getInputStream()));
+
+            if (pr.exitValue() != 0) {
+
+                while ((line = br.readLine()) != null) {
+                    System.out.println(line);
+                }
+
+                System.exit(1);
+            }
+
             pr.destroy();
 
             System.out.println("End of the psiTNorm.");
 
+            // Put the file with the header in the first position in the folder.
+            pt = new Path(psiT);
+            status = fs.listStatus(pt);
+            for (FileStatus stat : status) {
+
+                if (stat.getPath().toString().indexOf("part-") > -1) {
+                    br = new BufferedReader(new InputStreamReader(fs.open(stat.
+                            getPath())));
+
+                    line = br.readLine();
+                    if (line.indexOf("#") > -1) {
+                        fs.rename(stat.getPath(), new Path(stat.getPath().
+                                toString().replaceAll("part-", "-")));
+                        break;
+                    }
+                }
+
+            }
 
             // Merge psiT output files.
             pt = new Path(psiT);
@@ -268,30 +334,152 @@ public class QWS {
                     null);
             fs.rename(new Path(psiT + "_New"), pt);
 
-            // Delete the OUTPUT_DIR if it exists.
-            fu.fullyDelete(new File(OUTPUT_DIR));
+            // Start of the absSquare
+            absSquare = workDir + "absSquare";
+
+            // Delete the output directory if exists.
+            pt = new Path(absSquare);
+            fs.delete(pt, true);
+
+            /*
+             * Computes the square of the absolute value for each element of the
+             * array.
+             */
+            pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
+                    + "AbsSquare " + psiT + " " + absSquare);
+
+            pr.waitFor();
+
+            br = new BufferedReader(new InputStreamReader(
+                    pr.getInputStream()));
+
+            if (pr.exitValue() != 0) {
+
+                while ((line = br.readLine()) != null) {
+                    System.out.println(line);
+                }
+
+                System.exit(1);
+            }
+
+            pr.destroy();
+
+            // End of the absSquare
+
+            // Start of the reshape
+            reshape = workDir + "reshape";
+
+            // Delete the output directory if exists.
+            pt = new Path(reshape);
+            fs.delete(pt, true);
+
+            // Gives a new shape for the array.
+            pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
+                    + "Reshape " + newShape + " " + absSquare + " " + reshape);
+
+            pr.waitFor();
+
+            br = new BufferedReader(new InputStreamReader(
+                    pr.getInputStream()));
+
+            if (pr.exitValue() != 0) {
+
+                while ((line = br.readLine()) != null) {
+                    System.out.println(line);
+                }
+
+                System.exit(1);
+            }
+
+            pr.destroy();
+
+            // End of the reshape
+
+            /*
+             * Start of the sumAxis. In this case the output of SunAxis function
+             * will be the PDF of psiT.
+             */
+            pdf = workDir + "pdf";
+
+            // Delete the output directory if exists.
+            pt = new Path(pdf);
+            fs.delete(pt, true);
+
+            /*
+             * Sum the elements of the array over given axes, for a specific
+             * position.
+             */ 
+            pr = rt.exec("hadoop jar " + jarDir + "operations.jar operations."
+                    + "SumAxis " + axes + " " +  position + " " + reshape + " "
+                    + pdf);
+
+            pr.waitFor();
+
+            br = new BufferedReader(new InputStreamReader(
+                    pr.getInputStream()));
+
+            if (pr.exitValue() != 0) {
+
+                while ((line = br.readLine()) != null) {
+                    System.out.println(line);
+                }
+
+                System.exit(1);
+            }
+
+            pr.destroy();
+
+            // End of the sumAxis
+
+            System.out.println("End of the pdf.");
+
+            /*
+             * Delete _logs folder and _SUCCESS file in the psiT and psiTNorm
+             * folders.
+             */
+            pt = new Path(psiT + "/_logs");
+            fs.delete(pt, true);
+            pt = new Path(psiT + "/_SUCCESS");
+            fs.delete(pt, false);
+            pt = new Path(psiTNorm + "/_logs");
+            fs.delete(pt, true);
+            pt = new Path(psiTNorm + "/_SUCCESS");
+            fs.delete(pt, false);
+            pt = new Path(pdf + "/_logs");
+            fs.delete(pt, true);
+            pt = new Path(pdf + "/_SUCCESS");
+            fs.delete(pt, false);
+
+            // Delete the outputDir if it exists.
+            fu.fullyDelete(new File(outputDir));
 
             // Copy the result from HDFS to local.
             pt = new Path(psiT);
-            fu.copy(fs, pt, new File(OUTPUT_DIR + "psiT"), false, conf);
+            fu.copy(fs, pt, new File(outputDir + "psiT"), false, conf);
             pt = new Path(psiTNorm);
-            fu.copy(fs, pt, new File(OUTPUT_DIR + "psiTNorm"), false, conf);
+            fu.copy(fs, pt, new File(outputDir + "psiTNorm"), false, conf);
+            pt = new Path(pdf);
+            fu.copy(fs, pt, new File(outputDir + "pdf"), false, conf);
 
-
-            // Delete the WORK_DIR directory.
-            pt = new Path(WORK_DIR);
+            // Delete the workDir directory.
+            pt = new Path(workDir);
             fs.delete(pt, true);
 
             fs.close();
-
 
             System.out.println("Finished!");
 
             System.out.println("Runtime = " + ((System.nanoTime() - startTime)
                     / Math.pow(10, 9)) + " seconds");
 
+        } catch (NullPointerException e) {
+
+            System.out.println("Some input path has a empty directory.");
+
         } catch (Exception e) {
+
             System.out.println(e);
+
         }
 
     }
